# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 公司管理智能体 后端 API 契约
  version: 0.1.0
  description: |
    唯一接口事实来源。任何人新增/修改接口，必须先改本文件并经评审，再写代码。
    所有业务接口需携带 JWT；权限由后端中间件按 employees.permissions 校验。
servers:
  - url: http://localhost:8000
    description: 本地开发

security:
  - bearerAuth: []

paths:

  /api/auth/login:
    post:
      summary: 登录（身份识别）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emp_id, password]
              properties:
                emp_id:   { type: integer, example: 220401, description: 工号 }
                password: { type: string, example: "******" }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:       { type: string }
                  role:        { type: string, enum: [employee, manager], description: 由岗位层级判定 }
                  department:  { type: string }
                  position:    { type: string }
                  permissions: { type: array, items: { type: string } }
        "401": { description: 姓名或密码错误 }

  /api/attendance/checkin:
    post:
      summary: 打卡
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [clock_in, clock_out], default: clock_in }
      responses:
        "200":
          description: 打卡结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResult"

  /api/attendance/my:
    get:
      summary: 我的考勤记录
      parameters:
        - { name: period, in: query, schema: { type: string, enum: [week, month] }, required: true }
      responses:
        "200":
          description: 考勤列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  records: { type: array, items: { $ref: "#/components/schemas/Attendance" } }

  /api/attendance/report:
    get:
      summary: "[管理端] 考勤报表（按权限过滤）"
      parameters:
        - { name: dept,   in: query, schema: { type: string } }
        - { name: period, in: query, schema: { type: string, enum: [week, month] }, required: true }
      responses:
        "200":
          description: 报表数据 + Dify 生成的分析文案
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendanceReport" }

  /api/reimbursement/submit:
    post:
      summary: 提交报销
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:    { type: array, items: { type: string, format: binary } }
                category: { type: string }
                amount:   { type: number }
                desc:     { type: string }
      responses:
        "200":
          description: 返回缺项提醒与工单号
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_id: { type: integer }
                  missing:   { type: array, items: { type: string }, description: 缺失字段提示 }

  /api/reimbursement/my:
    get:
      summary: 我的报销进度
      responses:
        "200":
          description: 报销列表（含状态）
          content:
            application/json:
              schema:
                type: object
                properties:
                  records: { type: array, items: { $ref: "#/components/schemas/Reimbursement" } }

  /api/reimbursement/report:
    get:
      summary: "[管理端] 报销统计与异常分析"
      responses:
        "200":
          description: 统计明细 + Dify 生成的风险提示文案

  /api/salary/submit-performance:
    post:
      summary: 提交绩效达成情况，生成工资条
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [achievements]
              properties:
                achievements: { type: string, description: 绩效达成的自然语言描述 }
      responses:
        "200":
          description: 工资条
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Salary" }

  /api/salary/my:
    get:
      summary: 我的工资条列表
      responses:
        "200":
          description: 工资条列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  records: { type: array, items: { $ref: "#/components/schemas/Salary" } }

  /api/salary/report:
    get:
      summary: "[管理端] 工资分析"
      responses:
        "200":
          description: 工资构成与波动分析（Dify 文案）

  /api/assessment/query:
    post:
      summary: 查询岗位知识（内部调 Dify RAG）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties:
                position: { type: string, description: 想了解的岗位 }
      responses:
        "200":
          description: 岗位介绍/能力要求
          content:
            application/json:
              schema:
                type: object
                properties:
                  intro: { type: string }

  /api/assessment/stats:
    get:
      summary: "[管理端] 考核查询统计"
      responses:
        "200":
          description: 员工查询/答题情况统计

  /api/roster:
    get:
      summary: "[管理端] 花名册查询"
      responses:
        "200":
          description: 员工列表（不含密码哈希）
    post:
      summary: "[管理端] 新增员工"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201": { description: 已创建 }

  /api/roster/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      summary: "[管理端] 修改员工信息"
      responses:
        "200": { description: 已更新 }
    delete:
      summary: "[管理端] 删除员工（软删除待定）"
      responses:
        "204": { description: 已删除 }

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    CheckinResult:
      type: object
      properties:
        time:     { type: string, format: date-time }
        type:     { type: string, enum: [clock_in, clock_out] }
        is_late:  { type: boolean }
        message:  { type: string }

    Attendance:
      type: object
      properties:
        employee_id:  { type: integer }
        checkin_time: { type: string, format: date-time }
        type:         { type: string, enum: [clock_in, clock_out] }
        is_late:      { type: boolean }
        work_date:    { type: string, format: date }

    AttendanceReport:
      type: object
      properties:
        rows:   { type: array, items: { $ref: "#/components/schemas/Attendance" } }
        analysis: { type: string, description: Dify 生成的行为分析文案 }

    Reimbursement:
      type: object
      properties:
        employee_id: { type: integer }
        category:    { type: string }
        amount:      { type: number }
        status:      { type: string, enum: [draft, submitted, approving, approved, rejected] }
        approver_id: { type: integer, nullable: true }
        submit_time: { type: string, format: date-time }

    Salary:
      type: object
      properties:
        employee_id: { type: integer }
        period:      { type: string, example: "2026-08" }
        base:        { type: number }
        performance: { type: number }
        subsidy:     { type: number }
        total:       { type: number }

    EmployeeInput:
      type: object
      required: [emp_id, name, password, position, department]
      properties:
        no:          { type: integer }
        emp_id:      { type: integer, description: 工号，唯一登录账号 }
        name:        { type: string }
        password:    { type: string, description: 后端负责哈希，绝不入库明文 }
        permissions: { type: array, items: { type: string } }
        position:    { type: string }
        department:  { type: string }